Separate fundsTransferService and work on Frontend
