+++ modules/api/src/endpoints/user/merchantid.js
@@ -4,16 +4,9 @@
  * @file Defines a GET endpoint for /merchantid
  */
 import { mysql, handleResponse, auth } from 'utils'
-import url from 'url'
 
 let GET = async (req, res) => {
-  console.log('GET /merchantid requested')
-
-  // parse the provided data
-  const query = url.parse(req.url, true).query
-  console.log(query)
-
-  let userIndex = await auth(query.APIKey, res)
+  let userIndex = await auth(req.body.APIKey, res)
   if (!userIndex) return
 
   let result = await mysql.query(
