+++ modules/api/src/endpoints/user/sales.js
@@ -4,18 +4,10 @@
  * @file Defines a GET endpoint for /totalsales
  */
 import { mysql, handleResponse, auth } from 'utils'
-import url from 'url'
 
 let GET = async (req, res) => {
-  console.log('GET /totalsales requested')
-
-  // parse the provided data
-  const query = url.parse(req.url, true).query
-  console.log(query)
-
-  let userIndex = await auth(query.APIKey, res)
+  let userIndex = await auth(req.body.APIKey, res)
   if (!userIndex) return
-
   let result = await mysql.query(
     'SELECT totalSales FROM users WHERE tableIndex = ? LIMIT 1',
     [userIndex]
