+++ modules/website/src/API/index.js
@@ -7,4 +7,4 @@ export { default as merchantid  } from './merchantid'
 export { default as payments    } from './payments'
 export { default as setUsername } from './setUsername'
 export { default as setPassword } from './setPassword'
-export { default as newapikey   } from './newapikey'
+export { default as getapikeys  } from './getapikeys'
