+++ modules/website/src/API/utils/index.js
@@ -2,3 +2,4 @@ export { default as parseAddress  } from './parseAddress'
 export { default as generateError } from './generateError'
 export { default as get           } from './get'
 export { default as post          } from './post'
+export { default as patch         } from './patch'
