+++ modules/website/src/components/Payment/payment.css
@@ -10,3 +10,10 @@
 .payment-inner {
   margin: 1em;
 }
+
+.payment-transaction {
+  background-color: #ccccff;
+  overflow-y: hidden;
+  margin-bottom: 0.25em;
+  margin-top: 0.25em;
+}
