+++ modules/website/src/pages/portal/Settings/APIInfo.js
@@ -1,30 +1,30 @@
 import React, { Component } from 'react'
-import { newapikey } from 'API'
+import { getapikeys } from 'API'
 import { Button } from '@material-ui/core'
 import { Container, Text, SourceCode } from 'components'
 
 export default class APIInfo extends Component {
 
   state = {
-    APIKey: sessionStorage.gatewayAPIKey,
-    showKey: false
+    APIKeys: [],
+    showKeys: false
   }
 
   constructor(props) {
     super(props)
     this.handleSubmit = this.handleSubmit.bind(this)
+    getapikeys().then((response) => {
+      if (response.status === 'success') {
+        this.setState({
+          APIKeys: response.APIKeys
+        })
+      }
+    })
   }
 
   handleSubmit (e) {
     e.preventDefault()
-    newapikey().then((response) => {
-      if (response.status === 'success') {
-        this.setState({
-          APIKey: response.newAPIKey
-        })
-        sessionStorage.gatewayAPIKey = response.newAPIKey
-      }
-    })
+    // ...
   }
 
   render() {
@@ -36,19 +36,23 @@ export default class APIInfo extends Component {
           means you can build custom services and solutions that fit your needs
           without running your own servers.
         </Text>
-        <h3>Your API Key</h3>
+        <h3>Your API Keys</h3>
         <Text>
-          Your account's API key is being used behind the scenes whenever you
-          use gateway.cash. To learn more about the Gateway API,
+          Your account needs to have at least one API key to operate. If all
+          your keys get deleted, a new one is generated for you automatically
+          at login. To learn more about the Gateway API,
           go <a href="https://api.gateway.cash">here</a>.
+        </Text>
+        <Text>
+          Don't share your API keys with anyone else. With these keys, someone could change your password, steal your funds and hijack your account!
         </Text>
         {
           this.state.showKey ? (
             <>
               <SourceCode>
-                {this.state.APIKey}
+                {this.state.APIKeys.toString()}
               </SourceCode>
-              <h3>Generate New Key</h3>
+              <h3>Generate New Key (not working yet)</h3>
               <Text>
                 Generating a new API key will invalidate your current key and
                 will log out any devices and services which use your merchant
@@ -70,9 +74,9 @@ export default class APIInfo extends Component {
             <center>
               <Button
                 color="primary"
-                onClick={() => this.setState({showKey: true})}
+                onClick={() => this.setState({showKeys: true})}
               >
-                Show Key
+                Show API Keys
               </Button>
             </center>
           )
